feature/login: fix component
